Fix gen-special when an entity is commented out:
In the xorg application, mkfontdir is commented out, but we did not
check that because it was complicated with the shell. This was
leading to mkfontfir appearing in the packages to install, which
is not wanted. We use the information in the "cat" command to
eliminate the commented out package.